+++ src/collective/contact/mailaction/browser/actions/mail_utils.py
@@ -1,21 +1,17 @@
 # -*- coding: utf-8 -*-
-from collective.contact.mailaction.interfaces import \
-    ICollectiveContactMailactionTemplate
+from collective.contact.mailaction.interfaces import ICollectiveContactMailactionTemplate
 from email.MIMEMultipart import MIMEMultipart
 from email.MIMEText import MIMEText
 from email.Utils import formatdate
 from plone import api
 from plone.api.exc import InvalidParameterError
+from Products.CMFPlone.utils import safe_unicode
 from zope import component
 
 import email
-import formatter
-import htmllib
+import html2text
 import logging
-import StringIO
 
-
-PATTERN = '************/'
 
 logger = logging.getLogger('collective.contact.mailaction')
 
@@ -54,10 +50,6 @@ def create_email_body(html, text=None, headers=None):
     plain text version
     """
 
-    # Use DumbWriters word wrapping to ensure that no text line
-    # is longer than plain_text_maxcols characters.
-    plain_text_maxcols = 72
-
     try:
         encoding = api.portal.get_registry_record('plone.email_charset')
     except InvalidParameterError:
@@ -68,32 +60,22 @@ def create_email_body(html, text=None, headers=None):
     if text is None:
         # Produce an approximate textual rendering of the HTML string,
         # unless you have been given a better version as an argument
-        textout = StringIO.StringIO()
-        formtext = formatter.AbstractFormatter(
-            formatter.DumbWriter(
-                textout,
-                plain_text_maxcols
-            )
-        )
-        parser = htmllib.HTMLParser(formtext)
-        parser.feed(html)
-        parser.close()
 
-        # append the anchorlist at the bottom of a message
-        # to keep the message readable.
-        counter = 0
-        anchorlist = "\n\n" + ("-" * plain_text_maxcols) + "\n\n"
-        for item in parser.anchorlist:
-            counter += 1
-            anchorlist += "[%d] %s\n" % (counter, item)
-
-        text = textout.getvalue() + anchorlist
-        del textout, formtext, parser, anchorlist
-
-        # strip off style definitions
-        pos = text.find(PATTERN)
-        if pos != -1:
-            text = text[pos + len(PATTERN):]
+        # see tests/test_newsletterformatter for configuration options
+        parser = html2text.HTML2Text()
+        # for images, simply show their alt-text
+        parser.ignore_images = False
+        parser.images_to_alt = True
+        # skip links to named anchors
+        parser.skip_internal_links = True
+        # if link text and href is the same, simply show link
+        parser.use_automatic_links = True
+        # protect links from linebreaks - looks good for emails
+        parser.protect_links = True
+        # true to show links next to their text content
+        parser.inline_links = True
+        text = parser.handle(
+            safe_unicode(html)).replace('|  ', '').replace('|\n', '')
 
     else:
         text = text.encode(encoding)